Give a warning if SingleValueResult annotation is omitted
If a user set a parametrized type as a return type of method
and didn't set @SingleValueResult then generic information
is erased and JDBI will try to map only generic type.

The problem that the user doesn't have a clue about why his
query doesn't work. He will just get a mapper exception or
null.

In this this case we could give a warning to him that he
probably forgot to place an annotation to the method.

This is highly useful during development, when probability
of missing an annotation is very high.