Support for a positional binder in SQL Object queries
By default users of the SQLObject API must map method parameters to
query arguments by specifying @Bind annotation with a name of the
argument.

This is a standard and most reliable way. But in most cases it seems
as overkill because method parameter names are often present in the
same order as query arguments. We are forced to write boilerplate
code to glue them.

It would be great to have facility not to specify binding names, but
rely on parameter positions and let the library handle them.

Moreover, the Fluent API has this facility and support for it in
the SQLObject API will make migration from one API to another more
smooth.