Add SQL script lexer for semicolons and comments
Currently semicolons in a script are treated as statements
separators. Not all are. Semicolons in string literals are
just symbols. Also we don't handle comments, that are not
started in the begginning of the line;

This change adds a lexer, which matches comments and semicolons
in an SQL script. This lexer will:

* Correctly handle quoted comments and semicolons;
* Peplace the current matching with regular expressions;