Arbitrarily Paramaterized Sql Locators
Change sql "naming" to be (Object,Object...) based instead of a String.
This allows for much more sophisticated naming strategies, such as using
the SQL Object class and the method name, or doing classpath lookups
relative to the passed in class, etc.

In order to keep flexibility, it is done as just objects. Ideally it
would be type parameterized on the sql locator, but that would require
the sql locator setter become a builder, plus it would make for *really*
verbose type names, which would suck. This seems like a pragmatic
compromise, and fits with other dynamic aspects of jdbi.