Add a bridge for SqlObject API
In JDBI 2 there is the possibilty to create proxies for SqlObject interfaces
via the `DBI` and `Handle` class instances. Actually, it's very convienent and
even JDBI official docs mentions this as the standard way for using the SqlObject API.

This API was removed in JDBI, because the JDBI core and the SQLObject API are
now in different modules.

This commit adds a bridge from the JDBI core to the SQLObject API. It uses
method handles to dynamically call the methods of the `SqlObjectBuilder` class.

It changes some tests to show that the bridge to `SqlObjectBuilder` in the JDBI core actually works.

Also it fixes a broken link in the `SqlObjectBuilder` javadoc.