Allows binding bean of inaccessible type
Changes the BinderFactory interface so that the type token of the
annotated parameter is accepted as a parameter on #build.

Changes BindBeanFactory to get the value of its argument using the
Method of the annotated parameter's type instead of the argument's type.
This allows for BindBeanFactory to access methods on args that are of
types that are inaccessible to the SqlObject package.  This changes the
behavior of the BindBeanFactory to behave in a way that is much more
consistent with Java's polymorphic/access mechanism.

Using BindBean with an iterable of a raw type will result in using
Object.  This may be considered a limiting factor or a feature.

Fixes #242