Allows binding bean of inaccessible type
Adds a field to the BindBean annotation so that a type can be explicitly
set for the purpose of resolving bean bindings.  If specified this type
will be used instead of the provided argument's type.

The BindBean#type field allows for BindBeanFactory to access methods on
args that are of types that are inaccessible to the SqlObject package, but
are members of a supertype that is accessible to the package.

Fixes #242