Providing a list of column names returned in a prepared batch
Sometimes it's desirable have the ability to explicitly provide the names
of the columns, which should returned to the client after a query execution.
For example, when a row contains many columns, and many rows are returned in
a batch response.

JDBC provides such functionality during preparing a statement by
using form `Connection.preparedStatement(String query, String[] columnNames)`.

If a JDBC driver supports it, it will return only specified columns in a
response. For example, the official PostgreSQL driver adds a `RETURNING`
clause with a list of the specified columns to a query.

See this [SO discussion](http://stackoverflow.com/questions/34560647/pg-jdbc-appends-returning-on-prepared-batch-wtih-existing-returning-clause)