Move Maven Toolchains to a separate profile
Toolchains are useful when we compile against different JDKs or want to use a
specific old JDK. Since we target JDK8 in JDBI3, there is no big benefit in
using toolchains. In most development installations JDK8 is default and usually
already set in the PATH environment variable, so Maven can find it without
specific assistance. We only need to set-up source and target for the Java
compiler.

Also, avoiding using toolchains simplifies the process of contributing to JDBI,
because developers don't need to add specific steps to configure the development
environment and can build the project with mvn clean install.