Correctly handle null values in a response in JDBI3
Users often use methods which return only a single result.
In this case a returned `null` indicates that a RDBMS didn't
find a document which satisifies the passed arguments.
But actually the returned document can be `null` itself,
if it's a nullable column.

We should expect such values and return them as null/optional
values to the user. Currently we throw a `NullPointerException`
instead, which is a rather harsh strategy for handling nullable
values.