ClasspathSqlLocator: improve caching behavior to not depend on GC.
Current implementation uses a WeakHashMap -- this is bad for a cache
because the common case, in 'findSqlOnClasspath', uses a computed
key that doesn't exist outside the scope of the lookup.  So the moment
a lookup succeeds, the key is immediately thrown away, and is therefore
collected on the next minor GC.

This also makes the tests flaky I think.  I'm observing intermittent
failures on Travis.

Guava solves this nicely with the Cache class, configured to expire
after an interval of disuse rather than relying on GC.  I wish there
was a lighter weight solution but I didn't see any obvious choices
in the base JDK libraries.