Allow imperative SQL object methods to locate SQL statements in the same classpath directories as declarative methods.
Add methods Handle.getSqlObjectType, Handle.getSqlObjectMethod, and Handle.setSqlObjectContext (sets both
properties at once). These values are forwarded to the StatementContext of any Query, Update, Batch, etc
upon creation of a SQLStatement.

In v3, we may want to hoist the setting of these values to a universal Handler wrapper which gets wrapped
around every method. In this way we set the properties once on the Handle and would not have to expose
setters on the StatementContext.