Reuse a Postgres instance during tests
Starting up and shutting down a Postgres instance are rather heavy operations.
As a consequence of this, unit tests for the Postgres module takes a
considerable amount of time (1 min). We could speed this process up, by reusing
`PostgresDbRule` on the unit class level, and recreating the relevant table
before every test. This should be a more fast process (~15 sec), because
dropping and creating tables are relatively cheap operations.