Move tests to AssertJ
AssertJ is modern testing library which allows to write test assertions
which are easy to read and understand. It supports Java 8 features and
allows to perform assertions in the chain style.

This changes adopts AssertJ as JDBI's testing library and removes
all usages of Hamcrest assertions, which are fine, but are rather
limited.