Make `StatementContext` an interface
And create a concrete implementation of it, which is package-private
accessible for internal JDBI components. This allows to restrict the
amount of components which can change statement context's state, but
at the same time to keep a public interface which allows to read the
state. This also make statement contexts more testable because the
use code which interacts with statement contexts can mock the interface.