merge ConfigurerFactory into StatementCustomizers, rework parameter customization
Turns out the main reason for the ConfigurerFactory vs StatementCustomizer is
simply because StatementCustomizer isn't quite the right abstraction.

Move ConfigRegistry access up to Handler creation.  Pass it in to the customizer's
create* methods.  This allows you to mutate it once upfront, and choose whether or
not to customize the method further.

Fix createForParameter signature to make more sense

Fixes #647