Refactor HandlerFactory to allow for alternative method handler mapping strategies.
I'm submitting this as an alternative to PR #755 (which would add a config property for a "default" method handler as an alternative to existing default method handler).

However I believe the approach in this commit is a better choice long-term, as it allows clients to override the out-of-the-box behavior for any method,
rather than only being able to swap out the "fall-back" method handler.

The public API changes are modest:

* [BREAKING CHANGE] Change HandlerFactory to return Optional<Handler> instead of Handler
* [BREAKING CHANGE] Change semantics of `@SqlMethodAnnotation`: `value()` returns a handler type instead of handler factory type.
  * The handler type must have a public no-arg, (Method), or (Class, Method) constructor.
* Add Handlers registry with last-registered-wins semantics like our other registries
* Refactor SqlObjectFactory toi use the Handlers registry.
  * Existing logic was moved out to SqlMethodHandlerFactory and DefaultMethodHandlerFactory, which are registered out of the box.