Add a check for mistyped named parameters
It's very easy to mistype an SQL query and use a column name instead of
a naming parameters. For example, `DELETE FROM users WHERE id=id` instead
of `DELETE FROM users WHERE id=:id`. Unfortunatelly, this query is a valid
SQL query and a DBMS will happily execute it. Such mistake is very easy to
make, relatively hard to catch with a unit test and can lead to a disaster
in production.

It would be great if JDBI could detect situations when the user added
a binding, but the query doesn't have any named parameters.