Initial handshake
This change includes the initial handshake between the client and the
server. It is fixed to a username and password authentication
(although trust authentication would work as well) and only gives the
ability to execute a simple query. The r2dbc-api as it exists here is
only a placeholder proof-of-concept and shouldn’t be considered a
candidate for an actual API.