Refactor Framing
This change refactors what was previously called "framing".  It starts by
moving the message flow, of which there are only a small number into the core
package.  This co-locates then with the types that are going to use them.  It
also breaks out the authentication handlers (the types that know how to
respond to an authentication challenge) into an interface and a package that
will contain all implementations of them eventually.  Finally all of those
refactored parts are tested completely.  The TcpClient, the kernel of the
server interaction still isn't tested, but it should be next on the list.