Transaction Mode
This change adds transaction mode support to the API. It defines both
Isolation Level (e.g. READ COMMITTED) and Mutability (READ WRITE). An
API on Transaction allows users to set the transaction mode for a
specific, open transaction while and API of Connection allows users
to se the transaction mode for all transactions opened by a give
connection. In addition to the API’s the PostgreSQL implementation of
these APIs was created. Examples of both API usages were created, but
they are ignore as they’ve highlighted that the case where the server
sends an error breaks the framing of subsequent operations.