Cleanup Windowing
Previously the windowing in the system was a bit haphazard, sometimes
including the delimiter, other times not.  This change unifies the windowing
in the system, suppressing the delimiters.  In addition, this change ensures
that downstream consumers of the windows stop listening when they receive the
terminal message they expect, regardless of whether the window is complete. In
fact, this leads to the case where the windows is typically canceled before it
completes (even though the very next signal was likely to be complete).  This
exacerbates (by design) the problem where the incoming buffers are not drained
on cancellation.