Warnings Compliance
Previously warnings here handled, logged, and suppressed, within each window,
however according to the spec, warnings should be handled regardless of the
state of the connection (e.g. if it is idle). This change updates to the code
to handle, log, and suppress warnings as soon as they are decoded, rather
than waiting for an open connection window to exist. Errors are still handled
within the window to ensure that only the window fails, not the entire
connection.