Send Terminate Message on Close
Previously, when the connection was closed, the inbound and outbound message
queues were closed which caused the TCP socket to be closed as well. While
this works, and PostgreSQL can handle it, it's not an ideal way to terminate
the the connection. This change updates the PostgreSQL implementation to send
the Terminate message and then close everything out. Because of this, the
Connection API was changed, but to a signature that it arguably should always
Have been.