Cleanup Message Pipeline
Previously, the message receiver did quite a bit of extraneous configuration
in attempt to allow multiple, incremental subscription. It turns out that
none of this was necessary. The only required configuration was to ensure
that the processor didn't cancel itself when a subscriber canceled. This
change simplifies that code and ensures that this configuration is now made.
In addition, this code highlighted a bug in reactor-core so it updates the
dependency to the latest snapshot. This change should be removed once the fix
Has been released.