Better Windowing
Previously, windows were created using the convenience methods identifying
boundaries.  Unfortunately, this created empty windows at the end of Fluxes,
and while this isn't an issue in most cases, this was unacceptable in the case
of a database driver.  This change revamps the windowing support in order to
ensure that windows are only opened when there is more data to come.