Test Client Windows
Previously the TestClient didn't support windows as a regular client would.
This led to weird behavior including a bunch of workarounds as to when to
complete a Flux and when not to. This change throws all those workarounds
away and instead properly institutes windowing. While the full window
functionality is now available, there are convenience methods that will be
used most of the time that are equivalent to the previous API.