Restructure Client
Previously the Client existed in the same package with the implementations of
the major APIs.  This was preferable simply for visibility reasons but as the
number of message flows was expanding, it became clear that there was just too
much going on in that package.  This change restructures the code such that
the client and the message flows that use it move into their own isolated
package.  Unfortunately this means that a number of types become more visible
(and possibly abused), but pruning the documentation is likely to help there.