Asynchronous Operations
This change adds support for asynchronous operations, specifically handling of
ParameterStatus which may arrive at any time.  While adding this support it
became apparent that the handling of BackendKeyData and ParameterStatus was
happening at the wrong level abstraction.  That handling has now been moved
down into the client as they, like the ErrorResponse and NoticeResponse may
arrive at anytime in an exchange.