Simplify API
Previously, there was a bit of complication as a "convenience" API was
included in the "core" API that would need to be implemented by individual
databases. As the "convenience" API could be shared across all database
implementations, it makes more sense to move that out into a shared
implementation and keep the "core" API as lean as possible. As part of this,
most files moved around to new projects. In addition I'd consider the "core"
API more or less complete, with the caveat that it may change as more
Implementation drives changes.