Row Improvements
Previously the Result type would allow Rows to be used outside of its
lifecycle.  Unfortunately, this meant that the underlying implementation
(specifically Netty ByteBufs) could leak since there was no good way to know
when the data was no longer used.  This change updates the API such that, Rows
must be transformed within the lifecycle of the result giving the
implementation an opportunity to clean up after itself.

In addition this change adds decoding functionality for all the core types and
cleans up most of the existing ByteBuf leaks.  There is still a leak somewhere
to find, but it's good enough for now.