Fix memory leak in StringTemplateEngine
When an instance of `ST` is created without an explicit group, it adds
the parsed template to `STGroup.defaultGroup`. This leads to a leak
everytime a query is parsed, as it is a singleton.

This change passes a new instance of `STGroup` everytime a new instance
of `ST` is created and thus allows GC to reclaim the templates.