Fix for dropwizard-jdbi3 build issues.
DropWizard JDBI3 library relies on the enforcer plugin to ensure that
all dependencies use the same version (dependency convergence).
Unfortunately, due to the versions of antlr-runtime and springtemplate
used by JDBI3, this check fails.

In order to fix this issue, this commit features two changes to the
parent pom.xml:

1) Excluding the antlr-runtime library from the dependency declaration
for antlr's stringtemplete; the newer version of stringtemplate brings
in an older version of the runtime.

2) Enabling the enforcer plugin so that should such an issue happen
again, the JDBI build will fail before such an issue is released (and
thus breaking projects that depend on JDBI).