Improve JdbiFactoryBean
This commit improves the JdbiFactoryBean. The getObject() method
re-created the Jdbi instances each time getObject() was called.
It is better to create it once and cache it afterwards. This
could also potentially break the contract as Jdbi should be
a singleton (according to the isSingleton method).

To further improve it build upon the AbstractFactoryBean from
Spring itself which allows to specify if it should be a singleton
or not and act accordingly.