Fix misworded comment in Vavr Option example
When name is bound to Option.none() in the query `select * from users
where :name is null or name = :name`, the first clause of the `WHERE`
evaluates to `null is null` which is true, so it matches _any_ row, not
necessarily a row where the name `IS NULL`.

Not sure if this is what you were going for, though. If the intent was
to illustrate an example with the original comment's meaning, the query
should probably be changed to something like `select * from users where
name IS NOT DISTINCT FROM :name`