Support Immutables getter methods with @ColumnName
https://github.com/jdbi/jdbi/pull/1661 allows Immutables property to be annotated
with @ColumnName, but it requires the setter name to be either the exact getter
name, or "set" + column name. None of the two works for the following common scenario:

```
    @Value.Immutable
    public interface GetterWithColumnName {
        @ColumnName("some_foo_column")
        int getFoo();
    }
```

For the example above, the setter must be named "getFoo" or "setsome_foo_column",
both are kinda awkward and not directly supported by Immutables.

This PR addresses it by trying to extract the default Immutables setter name from
the getter name.

Also a couple of minor cleanups, such as removing the obsolete SuppressWarnings.