Remove antlr4-runtime dependency
Inline the dependency into the core jar. This removes the long
standing problem that jdbi creates version conflicts when used in
projects that also use antlr4 and that use a different runtime
version. Unfortunately antlr insists on a strong relationship between
the generated classes for the grammar and the runtime.

Inlining this dependency makes allows JDBI users to choose their antlr
version arbitrarily as JDBI will use its private version.