Check connection status in handle.close()
Ensure that the handle will tolerate a closed connection when close()
is called. If the database connection gets closed between the last
operation finishing successfully and the handle being closed, checking
the autocommit status will cause an exception to be thrown that is
user-visible but not actionable. Check the connection state before
executing any operation on the database connection.

Fixes #2065