Javadoc updates
Clarify that the various methods that take strings are not deprecated
but it is preferred to use the CharSequence equivalents if possible.

We will *never* remove the String methods as rewriting existing code
to use the CharSequence methods would imply that every call with a
simple String either needs to be cast to CharSequence or wrapped into
Sql (which is still in beta). Either option is very questionable.

Restore the docs makes them useful when popping up in an IDE.