fix(generator): return latest supported SourceVersion instead of 8
Fixes #2127.

We should be able to output generated code for any supported releases >= 8.
Considering the bytecode we output for the artifact (i.e. the processor,
not the generated code) is always Java 8 or newer as of writing, this will
not be an issue.

We return the latest supported version for javac instead of the lowest
version we expect to support, because javac will print warnings in that case.
This breaks -Werror builds on versions higher than this is made to support
(i.e. Java 9+ as of writing).

Legal note: this code is submitted under the applicable licence(s)
(Apache 2.0, Jan 2004) with the ownership retained as Spotify AB.