Use native pg on macos arm64
The default postgres version (13) does not have a native arm64 build
for modern macos. Select a version that is supported so that native
code is run (and not rosetta translated x86_64).

Building on a M<x> mac may also need to apply the changes to SHMMAX
described here: https://dansketcher.com/2021/03/30/shmmax-error-on-big-sur/