Iteration on the MethodHandle.invokeExact code
Move all of the error handling into JdbiClassUtils, all places that
need to create an instance can do so by calling handle.invokeExact
with the right parameters as a lambda on all call sites.

This looks much cleaner code-wise but probably need some benchmarking
to ensure that this does not lose performance.