deprecated otj-postgres support
The OtjPostgres project ships its artifact as an automatic module with
filename based module name. It is strongly discouraged to ship any JPMS
enabled project that depends on automatic modules with filename based
module names. We have reached out to the OtjPostgres project and
proposed a change that would fix this.

If they ship a version that fixes this before we are ready to move to
full JPMS support, we will un-deprecate this class, otherwise it will be
removed with our switch to JPMS.