Add support for function arguments
Similar to `Consumer<...>` arguments, a function argument consumes the
results of a SQL query. The return value from applying the function is
returned by the SQL object method. This is especially useful for
collecting and managing streaming data as the function is called while
the database connection is live.

Addresses #2326