Correctness fix for Handle
There is a really obscure (and usually already fatal) error condition
where during handle creation, the connection object becomes no longer
viable (e.g. the database crashes, the connection gets closed). In that
case, either `transactionHandler.specialize()` or
`transactionHandler.isInTransaction()` will throw a SQLException (or a
runtime exception) and the cleanup handler for the connection is never
added. Technically, the handle will there leak the connection (which is
already non-viable because it caused the exception in the first
place). This fix moves the cleanup before calling the methods which will
ensure that the connection will be cleaned up in every situation.

This may change user visible behavior slightly; instead of a
SQLException, it is possible that now a `ConnectionException` with the
root exception attached is now observed. This is unlikely but possible.