Make trailing semicolons in scripts configurable
Turns out that MySQL and Oracle simply have differing ideas on what a
statement in a script should be. Oracle needs the trailing semicolons,
MySQL throws syntax errors (when using the
rewriteBatchedStatements=true setting, otherwise it reports a much
saner "cannot issue select statements with update" type of error
message.

Added a switch that retains the default behavior but can be flipped
to allow the error case described in #2554 to work.

Reported by @IrinaTerlizhenko, thank you for this!