Support null for bean, pojo, field, and method binding
Instead of throwing a NPE, do nothing. This allows for calling
the various Jdbi methods (and the annotations for sqlobject) to deal
with null objects.

While this is a behavior change, it is acceptable as callers had to
wrap this in `if (bean != null) { ...` conditionals and had to skip
binding for null values. Now the code does skip null values directly.