Clarify handle uses for sql objects
modify the configuration for a sql object by calling
`getHandle().register...` from the sql object. This does not work
because of the inner function of the extension framework.

Add a note to the documentation and put the various tests that came
out of the investigation into the integration testing suite.