Support a result set for Call
Some database drivers do not use out parameters but an actual result
set to return data from stored procedures. Expose the result set
through the `OutParameters` object.

Unfortunately, this means that it is no longer possible to close the
`CallableStatement` at the end of the operation but it now must be
resource managed unless the ResultSet is used (and iterated all the
way to the end).