Fix invoking subclass methods with on-demand extensions
Right now we build a Map with Method keys. The Method refers to the
declaring type, but we don't actually care if it is some super or subtype,
we just want to call the handler either way. We already know there's only
one possible entry due to checks in the extension metadata builder

Fixes #2769